requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cleanup==8.1.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
phonenumbers==8.13.33
pillow==10.3.0
sqlparse==0.4.4
